Connects #97:  Fixes an issue when checking employee local times against the time a scheduled message should be sent.
This changeset addresses an issue that could potentially prevent a message from being sent if the local time of a user does not exactly match (by the hour and minute) the time a scheduled message is set to be sent out.  This adjustment changes the check to look for a match at the current time or any time thereafter.  A separate check is already in place that ensures the message has not already been sent to the employee.