Connects #100:  Fixes issue with messages sent at the wrong time.
The changeset introduced to fix issue #97 had another problem with it in that the time check between the current time of the employee and the time the scheduled message was supposed to go out was still incorrect.  This changeset fixes the time comparison check and bolsters the tests to account for this situation.