Connects #104:  Add diagnostic page for time and time zone info.
This changeset adds a diagnostics page that displays application and server information regarding the current time and time zone settings.  It also slightly modifies the time check for matching messages to employees to be a bit more explicit and adds a Clockwork default time zone configuration for UTC.  The hope is that with these changes and additional information being exposed we can finally uncover why the scheduled messages have been sent out at the wrong time.