Readme improvement: more info about contributing
The CONTRIBUTING.md file for this project has lots of super helpful information about setup and finding tasks in addition to pull request guidelines, so I suggest slightly expanding the readme to encourage people to read the contributing file for that information as well, to help people find it. Here's a small suggestion - feel free to modify. :)