Rename and refactor match_time
**Why**:
Renaming `match_time` to `time_to_send_message?` better expresses what
it does, which is to determine whether or not it's time to send a
message. And to answer that question consistently with either true or
false (as opposed to true or nil), we use a guard clause at the top to
make it clear what the failing scenario is.

If a method returns a boolean, the convention is to make it end with a
question mark.