Fix issue where east coast users didn't get messages after 4:00 PM
Prior to this commit, messages would not go to users after a certain time depending on their timezone.
This is because the of an issues with the day advancing in the UTC timezone before advancing in the user's timezone.

The app would compare the hours, minutes, and day to determine if a message should be sent.
The UTC offset for PT is 8 hours (7 timezones and an hour for DST).
2016/9/15 16:00 PT is 2016/9/16 00:00:00 UTC.
When the app was comparing the days for messages after 4:00 PM, it was comparing 1 and 2 for the days and did not send a message since they were not equal.

This commit fixes the issue by refactoring the `time_to_sent?` method.

Instead of comparing the hours, minutes, and dates, individually, it now calculates the time that the message should be sent to the user and compares it to the current time.
If the current time is after the send time, it returns true.

This commit also updates a spec that was failing as a result of this change.
The spec that checks that the service class does not inappropriately send messages after midnight was failing with this comparison.

The spec was checking that the spec would not send a message scheduled for the evening of the previous day at the start of the current day.
This is erroneous because the app is built to send messages scheduled to be sent in the past if they have not been sent already.

This commit fixes the spec to move the scheduled day for the message to the current day.

closes #111