Refactor MessageSender to send general messages
This commit refactors the message sender and updates the specs so the message sender can better handle sending all 3 messages types:

- OnboardingMessages
- QuarterlyMessages
- BroadcastMessages

The message sender now creates a SentMessage for each type.
The spec also uses a shared example to run with each type of message.

This commit also moves some of the logic for the slack channel finder into a helper.