Remove Employee uniqueness validation from SentMessage
Ref #250

The SentMessage model had a uniqueness validation scoped to the message. The effect of this validation is that a SentMessage can only be created once for a given message and employee. As a result, sending a message to an employee twice resulted in a failure to create a SentMessage record for the second message.

This commit fixes the above bug by removing the validation.

A byproduct of this change is that if an employee has been sent onboarding message as part of a test before they were supposed to receive that message, they will not receive it at the scheduled time. This does not seem like a major concern since the employee will have ultimately still received the message.

Another byproduct is that an employee that has been sent a test quarterly message within a week of the beginning of a new quarter will not receive the message at the start of the new quarter. This does not seems like a concern since the employee will ultimately receive still receive the message near the beginning of the new quarter.

Finally this fix makes the assumption that there is not a reason to avoid creating SentMessage records for test messages. I am not aware of one, but if one exists, a follow up commit may be needed to tell the MessageSender not to create a SentMessage record for test messages. I think the fix on this commit still applies in that case, since we will still want to keep a record of when non-test messages get sent, even if they are sent twice for some reason.