Don't create SentMessage records for test messages
This commit adds an option to the MessageSender that tells the message it is sending is a test message. When this option is set to true, the message is sent, but a SentMessage record is not created.

This change allows users to receive onboarding messages after having been sent a test message for the given onboarding message. Prior to this commit, that did not happen because the logic to determine if an onboarding message should be sent to a user included a check for an existing onboarding message.

This commit also adds a feature spec to serve as a regression test to assure that users receive onboarding messages after receiving a test message for the given onboarding message.

Ref #250